Moved app.js files
